fix: don't throw client access error for env vars with empty string value (#10)
we should not throw an `InvalidPropertyAccessError` when trying to
access an optional env var with an empty string value.

example:

```
# .env.local
NEXT_PUBLIC_SOMETHING_OPTIONAL=
```

previously `NEXT_PUBLIC_SOMETHING` was deleted from `environment`
because its value was an empty string, thus throwing
`InvalidPropertyAccessError` when being accesed client-side.

we should treat empty string and undefined the same - which should fix
this. added a test as well.